// frontend/src/app/api/generated/model/verificationFactor.ts
/**
 * PulseDesk API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export enum VerificationFactor {
  DATE_OF_BIRTH = 'DATE_OF_BIRTH',

  POSTAL_CODE = 'POSTAL_CODE',

  ADDRESS = 'ADDRESS',

  PHONE = 'PHONE',
}